Move OxqlMetric files to own component directory
